+++ src/Expression/Builder/AbstractArrayExpressionBuilder.php
@@ -2,8 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Db\Expression;
+namespace Yiisoft\Db\Expression\Builder;
 
+use Yiisoft\Db\Expression\ArrayExpression;
+use Yiisoft\Db\Expression\Builder\ExpressionBuilderInterface;
+use Yiisoft\Db\Expression\ExpressionInterface;
 use Yiisoft\Db\Query\QueryInterface;
 use Yiisoft\Db\QueryBuilder\QueryBuilderInterface;
 use Yiisoft\Db\Schema\Data\LazyArrayInterface;