+++ src/Expression/Builder/ValueBuilder.php
@@ -2,8 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Db\Expression;
+namespace Yiisoft\Db\Expression\Builder;
 
+use Yiisoft\Db\Expression\Builder\ExpressionBuilderInterface;
+use Yiisoft\Db\Expression\ExpressionInterface;
+use Yiisoft\Db\Expression\Value;
 use Yiisoft\Db\QueryBuilder\QueryBuilderInterface;
 
 /**